Maluff

+++ backend/main/resources/opinion.py
@@ -3,8 +3,12 @@ from flask import request, jsonify
 from main.models import OpinionModel, PrestamoModel
 from .exception import IdEnUso
 from .. import db
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from main.auth.decorators import role_required
 
 class Opinion(Resource):
+    
+    @jwt_required(optional=True)
     def get(self, id):
         try:
             opinion = db.session.query(OpinionModel).get_or_404(id)
@@ -12,17 +16,20 @@ class Opinion(Resource):
         except Exception:
             abort(500, message=str("Error 404: el id de la opinion no existe"))
     
+    @role_required(["admin"])
     def delete(self, id):
         try:
             opinion = db.session.query(OpinionModel).get_or_404(id)
-            db.session.delete(opinion)
-            db.session.commit()
-            return 'La opinion fue borrada de manera satisfactoria', 201
+            current_identity = get_jwt_identity()
+            if current_identity == id:
+                db.session.delete(opinion)
+                db.session.commit()
+                return 'La opinion fue borrada de manera satisfactoria', 201
         except Exception as e:
             db.session.rollback()
             abort(500, message=str("404 Not Found: No se encuentra la opinion para eliminar. El ID no existe"))
     
-    
+    @role_required(['user'])
     def put(self, id):
         try:
             opinion = db.session.query(OpinionModel).get_or_404(id)
@@ -39,6 +46,7 @@ class Opinion(Resource):
 
 
 class Opiniones(Resource):
+    @jwt_required(optional=True)
     def get(self):
         page = 1
         per_page = 5
@@ -62,7 +70,7 @@ class Opiniones(Resource):
                   'page': page
                 })
         
-
+    @role_required(["user"])
     def post(self):
         data = request.get_json()
         if isinstance(data, dict):
