+++ backend/main/resources/usuario.py
@@ -5,26 +5,49 @@ from .. import db
 from flask import jsonify
 class IdEnUso(Exception):
     ...
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from main.auth.decorators import role_required
+from sqlalchemy import func, desc
+from .. import jwt
+from flask_jwt_extended import verify_jwt_in_request, get_jwt
 
 class Usuario(Resource):
+    @jwt_required(optional=True) #USUARIO ACCEDE AL GET, ADMINISTRADOR TAMBIEN PERO INFO MAS REDUCIDA. Pero opcional porque no logueado tambien puede ver
     def get(self, id):
-        try:
+        self.id = id
+        try:          
             usuario = db.session.query(UsuarioModel).get_or_404(id)
-            return usuario.to_json()
+            current_identity = get_jwt_identity() #get_jwt_identity() es el id del token que sera el del usuario
+           
+            #Comparo el id que se pide con el id perteneciente al token
+            if current_identity == id:                    
+                return usuario.to_json()
+            else:
+                return usuario.to_json_short() #Si no existe token, mostrar solo datos relevantes
         except Exception:
             abort(404, message=str("Error 404: el id del usuario no existe"))
-
-
+    
+    
+    @role_required(roles = ["user", "admin"]) #En token viene un rol que debe ser alguno de los dos, para poder borrar
     def delete(self, id):
-        try:
+        self.id = id 
+        try: 
             usuario = db.session.query(UsuarioModel).get_or_404(id)
-            db.session.delete(usuario)
-            db.session.commit()
-            return 'El usuario fue borrado de manera satisfactoria', 201
+            current_identity = get_jwt_identity()
+            #Obtener claims de adentro del JWT
+            claims = get_jwt()
+            if claims["rol"]== "admin" or current_identity == id:
+                db.session.delete(usuario)
+                db.session.commit()
+                return 'El usuario fue borrado de manera satisfactoria', 201
+            else: 
+                return "Usted no posee la cuenta que quiere borrar", 404
         except Exception as e:
             db.session.rollback()
             abort(404, message=str("404 Not Found: No se encuentra el usuario para eliminar. El ID no existe"))
     
+    
+    @jwt_required(optional=True)    
     def put(self, id):
         try:
             usuario = db.session.query(UsuarioModel).get_or_404(id)
@@ -38,7 +61,10 @@ class Usuario(Resource):
             db.session.rollback()
             abort(404, message=str("Error 404 NOt Found: No se encuentra el usuario para modificar"))
 
+
+
 class Usuarios(Resource):
+    @role_required(roles = ["admin"])    
     def get(self):
         page = 1
         per_page = 10
@@ -71,32 +97,3 @@ class Usuarios(Resource):
                   'pages': usuarios.pages,
                   'page': page
                 })
-             
-
-    def post(self):
-        data = request.get_json()
-        if isinstance(data, dict):
-            data = [data]
-        usuarios_list = []
-        for usuario_data in data:
-            usuarios = UsuarioModel.from_json(usuario_data)
-            try:
-                tabla = UsuarioModel.query.all()
-                self.verificacion(usuario_data, tabla)
-            except Exception as e:
-                return {'error': str(e)}, 403
-            db.session.add(usuarios)
-            usuarios_list.append(usuarios)
-        db.session.commit()
-        usuario_json = [usuario.to_json() for usuario in usuarios_list]
-        return usuario_json, 201
-
-
-    def verificacion(self, usuario, tabla):
-        for i in tabla:
-            id = i.id
-            id_nuevo = usuario['id']
-            if id == id_nuevo:
-                raise IdEnUso('El ID esta en uso')
-            else:
-                return None
